Switch everything to floats

+++ src/gpucb.c
@@ -15,14 +15,14 @@ void initialize(const int I, const int N){
     
     I_ = I;
     N_ = N;
-    T_ = (double *) malloc(I * sizeof(double));
+    T_ = (float *) malloc(I * sizeof(float));
     X_ = (int *) malloc(2 * I * sizeof(int));
-    X_grid_ = (double *) malloc(2 * N * N * sizeof(double));
+    X_grid_ = (float *) malloc(2 * N * N * sizeof(float));
     sampled_ = (bool *) malloc(N * N * sizeof(bool));
-    mu_ = (double *) malloc(N * N * sizeof(double));
-    sigma_ = (double *) malloc(N * N * sizeof(double));
-    K_ = (double *) malloc(I * I * sizeof(double));
-    L_ = (double *) malloc(I * I * sizeof(double));
+    mu_ = (float *) malloc(N * N * sizeof(float));
+    sigma_ = (float *) malloc(N * N * sizeof(float));
+    K_ = (float *) malloc(I * I * sizeof(float));
+    L_ = (float *) malloc(I * I * sizeof(float));
 
     // Initialize matrices
     for (int i = 0; i < N*N; i++) {
@@ -38,10 +38,10 @@ void initialize(const int I, const int N){
     initialize_meshgrid_baseline(X_grid_, N_, GRID_MIN_, GRID_INC_);
 }
 
-void initialize_meshgrid_baseline(double *X_grid, int n, double min, double inc) {
-    double x = min;
+void initialize_meshgrid_baseline(float *X_grid, int n, float min, float inc) {
+    float x = min;
     for (int i = 0; i < n; i++) {
-        double y = min;
+        float y = min;
         for (int j = 0; j < n; j++) {
             X_grid[i * 2 * n + 2 * j] = y;
             X_grid[i * 2 * n + 2 * j + 1] = x; // With this assignment, meshgrid is the same as python code
@@ -51,41 +51,34 @@ void initialize_meshgrid_baseline(double *X_grid, int n, double min, double inc)
     }
 }
 
-double function_baseline(double x, double y) {
-    // double t = sin(x) + cos(y);
-    double t = -pow(x, 2) - pow(y, 2);
-    printf("(C code) Sampled: [%.2lf %.2lf] result %lf \n", x, y, t);
+float function_baseline(float x, float y) {
+    // float t = sin(x) + cos(y);
+    float t = -powf(x, 2) - powf(y, 2);
+    printf("(C code) Sampled: [%.2f %.2f] result %f \n", x, y, t);
     return t;
 }
 
-void learn_baseline(double *X_grid,
-           double *K,
-           double *L_T,
+void learn_baseline(float *X_grid,
+           float *K,
+           float *L_T,
            bool *sampled,
            int *X,
-           double *T,
+           float *T,
            int t,
            int maxIter,
-           double *mu,
-           double *sigma,
-           double(*kernel)(double *, double *, double *, double *),
-           const double beta,
+           float *mu,
+           float *sigma,
+           float(*kernel)(float *, float *, float *, float *),
+           const float beta,
            int n) {
-    /*
-     * grid_idx = self.argmax_ucb()
-    *  self.sample(self.X_grid[grid_idx])
-    *  for every point x:
-     *  gp_regression()
-    *  gp.fit(self.X, self.T)
-    *  mu1 = self.mu
-     */
+
     bool debug = true;
     int maxI = 0;
     int maxJ = 0;
-    double max = mu[0] + sqrt(beta) * sigma[0];
+    float max = mu[0] + sqrtf(beta) * sigma[0];
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
-            double currentValue = mu[i * n + j] + sqrt(beta) * sigma[i * n + j];
+            float currentValue = mu[i * n + j] + sqrtf(beta) * sigma[i * n + j];
 
             if (!sampled[i * n + j] && (currentValue > max)) {
                 max = currentValue;
@@ -102,10 +95,10 @@ void learn_baseline(double *X_grid,
     gp_regression_baseline(X_grid, K, L_T, X, T, t, maxIter, kernel, mu, sigma, n); // updating mu and sigma for every x in X_grid
 }
 
-double kernel2_baseline(double *x1, double *y1, double *x2, double *y2) {
+float kernel2_baseline(float *x1, float *y1, float *x2, float *y2) {
     // RBF kernel
-    double sigma = 1;
-    return exp(-((*x1 - *x2) * (*x1 - *x2) + (*y1 - *y2) * (*y1 - *y2)) / (2 * sigma * sigma));
+    float sigma = 1;
+    return expf(-((*x1 - *x2) * (*x1 - *x2) + (*y1 - *y2) * (*y1 - *y2)) / (2 * sigma * sigma));
 }
 
 void run(){
@@ -134,7 +127,7 @@ void clean(){
     n:    The size of the data in matrix A to decompose
     size: The actual size of the rows
  */
-void cholesky_baseline(double *A, int n, int size) {
+void cholesky_baseline(float *A, int n, int size) {
     for (int i = 0; i < n; ++i) {
 
         // Update the off diagonal entries first.
@@ -149,7 +142,7 @@ void cholesky_baseline(double *A, int n, int size) {
         for (int k = 0; k < i; ++k) {
             A[size * i + i] -= A[size * i + k] * A[size * i + k];
         }
-        A[size * i + i] = sqrt(A[size * i + i]);
+        A[size * i + i] = sqrtf(A[size * i + i]);
     }
 }
 
@@ -164,7 +157,7 @@ void cholesky_baseline(double *A, int n, int size) {
     size: The actual size of the rows
 
  */
-void incremental_cholesky_baseline(double *A, double *A_T, int n1, int n2, int size) {
+void incremental_cholesky_baseline(float *A, float *A_T, int n1, int n2, int size) {
     for (int i = n1; i < n2; ++i) {
         // Update the off diagonal entries.
         for (int j = 0; j < i; ++j) {
@@ -178,7 +171,7 @@ void incremental_cholesky_baseline(double *A, double *A_T, int n1, int n2, int s
         for (int k = 0; k < i; ++k) {
             A[size * i + i] -= A[size * i + k] * A[size * i + k];
         }
-        A[size * i + i] = sqrt(A[size * i + i]);
+        A[size * i + i] = sqrtf(A[size * i + i]);
         A_T[size*i + i] = A[size * i + i];
     }
 }
@@ -194,10 +187,10 @@ void incremental_cholesky_baseline(double *A, double *A_T, int n1, int n2, int s
  *      x: vector to put result in
  *      lower: if one the lower triangle system is solved, else the upper triangle system is solved.
 */
-void cholesky_solve2_baseline(int d, int size, double *LU, double *b, double *x, int lower) {
+void cholesky_solve2_baseline(int d, int size, float *LU, float *b, float *x, int lower) {
     if (lower == 1) {
         for (int i = 0; i < d; ++i) {
-            double sum = 0.;
+            float sum = 0.;
             for (int k = 0; k < i; ++k) {
                 sum += LU[i * size + k] * x[k];
             }
@@ -205,7 +198,7 @@ void cholesky_solve2_baseline(int d, int size, double *LU, double *b, double *x,
         }
     } else {
         for (int i = d - 1; i >= 0; --i) {
-            double sum = 0.;
+            float sum = 0.;
             for (int k = i + 1; k < d; ++k) {
                 sum += LU[i * size + k] * x[k];
             }
@@ -216,22 +209,22 @@ void cholesky_solve2_baseline(int d, int size, double *LU, double *b, double *x,
 }
 
 // Old version.
-void cholesky_solve_baseline(int d, double *LU, double *b, double *x) {
-    double y[d];
+void cholesky_solve_baseline(int d, float *LU, float *b, float *x) {
+    float y[d];
     for (int i = 0; i < d; ++i) {
-        double sum = 0.;
+        float sum = 0.;
         for (int k = 0; k < i; ++k)sum += LU[i * d + k] * y[k];
         y[i] = (b[i] - sum) / LU[i * d + i];
     }
     for (int i = d - 1; i >= 0; --i) {
-        double sum = 0.;
+        float sum = 0.;
         for (int k = i + 1; k < d; ++k)sum += LU[k * d + i] * x[k];
         x[i] = (y[i] - sum) / LU[i * d + i];
     }
 }
 
 
-void transpose_baseline(double *M, double *M_T, int d, int size) {
+void transpose_baseline(float *M, float *M_T, int d, int size) {
     for (int i = 0; i < d; ++i) {
         for (int j = 0; j < d; ++j) {
             M_T[j * size + i] = M[i * size + j];
@@ -240,16 +233,16 @@ void transpose_baseline(double *M, double *M_T, int d, int size) {
 }
 
 
-void gp_regression_baseline(double *X_grid,
-                            double *K,
-                            double *L_T,
+void gp_regression_baseline(float *X_grid,
+                            float *K,
+                            float *L_T,
                             int *X,
-                            double *T,
+                            float *T,
                             int t,
                             int maxIter,
-                            double   (*kernel)(double *, double *, double *, double *),
-                            double *mu,
-                            double *sigma,
+                            float   (*kernel)(float *, float *, float *, float *),
+                            float *mu,
+                            float *sigma,
                             int n) {
     int t_gp = t + 1;
 
@@ -274,9 +267,9 @@ void gp_regression_baseline(double *X_grid,
     incremental_cholesky_baseline(K, L_T, t_gp - 1, t_gp, maxIter);
 
     // 3. Compute alpha
-    double x[t_gp];
-    double alpha[t_gp];
-    double v[t_gp];
+    float x[t_gp];
+    float alpha[t_gp];
+    float v[t_gp];
 
 
     cholesky_solve2_baseline(t_gp, maxIter, K, T, x, 1);
@@ -288,33 +281,29 @@ void gp_regression_baseline(double *X_grid,
     {
         for (int j = 0; j < n; j++) // for all points in X_grid ([i][j])
         {
-            double x_star = X_grid[2 * n * i + 2 * j]; // Current grid point that we are looking at
-            double y_star = X_grid[2 * n * i + 2 * j + 1];
-            double k_star[t_gp];
+            float x_star = X_grid[2 * n * i + 2 * j]; // Current grid point that we are looking at
+            float y_star = X_grid[2 * n * i + 2 * j + 1];
+            float k_star[t_gp];
 
             for (int k = 0; k < t_gp; k++) {
                 int x_ = X[2 * k];
                 int y_ = X[2 * k + 1];
-                double arg1x = X_grid[x_ * 2 * n + 2 * y_];
-                double arg1y = X_grid[x_ * 2 * n + 2 * y_ + 1];
+                float arg1x = X_grid[x_ * 2 * n + 2 * y_];
+                float arg1y = X_grid[x_ * 2 * n + 2 * y_ + 1];
                 k_star[k] = (*kernel)(&arg1x, &arg1y, &x_star, &y_star);
             }
 
-            double f_star = 0;
+            float f_star = 0;
             for (int k = 0; k < t_gp; k++) {
                 //f_star += k_star[k] * alpha->data[k];
                 f_star += k_star[k] * alpha[k];
             }
 
             mu[i * n + j] = f_star;
-            //printf("fstar is: %lf", f_star);
-            //printf("write in mu at %d \n", i*n+j);
             cholesky_solve2_baseline(t_gp, maxIter, K, k_star, v, 1);
-            //printf("loop solve done\n");
 
-            double variance = (*kernel)(&x_star, &y_star, &x_star, &y_star);
+            float variance = (*kernel)(&x_star, &y_star, &x_star, &y_star);
             for (int k = 0; k < t_gp; k++) {
-                //variance -= v->data[k] * v->data[k];
                 variance -= v[k] * v[k];
             }
