+++ src/gpucb5.c
@@ -11,7 +11,7 @@
 const char *tag[10] = {"inlined"};
 
 void initialize(const int I, const int N) {
-    if(N % 8 !=0){
+    if (N % 8 != 0) {
         printf("n is not divisible by 8 !!! \n");
     }
 
@@ -21,14 +21,14 @@ void initialize(const int I, const int N) {
 
     I_ = I;
     N_ = N;
-    T_ = (double *) malloc(I * sizeof(double));
+    T_ = (float *) malloc(I * sizeof(float));
     X_ = (int *) malloc(2 * I * sizeof(int));
-    X_grid_ = (double *) malloc(2 * N * N * sizeof(double));
+    X_grid_ = (float *) malloc(2 * N * N * sizeof(float));
     sampled_ = (bool *) malloc(N * N * sizeof(bool));
-    mu_ = (double *) malloc(N * N * sizeof(double));
-    sigma_ = (double *) malloc(N * N * sizeof(double));
-    K_ = (double *) malloc(I * I * sizeof(double));
-    L_ = (double *) malloc(I * I * sizeof(double));
+    mu_ = (float *) malloc(N * N * sizeof(float));
+    sigma_ = (float *) malloc(N * N * sizeof(float));
+    K_ = (float *) malloc(I * I * sizeof(float));
+    L_ = (float *) malloc(I * I * sizeof(float));
 
     // Initialize matrices
     for (int i = 0; i < N * N; i++) {
@@ -44,10 +44,10 @@ void initialize(const int I, const int N) {
     initialize_meshgrid(X_grid_, N_, GRID_MIN_, GRID_INC_);
 }
 
-void initialize_meshgrid(double *X_grid, int n, double min, double inc) {
-    double x = min;
+void initialize_meshgrid(float *X_grid, int n, float min, float inc) {
+    float x = min;
     for (int i = 0; i < n; i++) {
-        double y = min;
+        float y = min;
         for (int j = 0; j < n; j++) {
             X_grid[i * 2 * n + 2 * j] = y;
             X_grid[i * 2 * n + 2 * j + 1] = x; // With this assignment, meshgrid is the same as python code
@@ -57,41 +57,33 @@ void initialize_meshgrid(double *X_grid, int n, double min, double inc) {
     }
 }
 
-double function(double x, double y) {
-    // double t = sin(x) + cos(y);
-    double t = -pow(x, 2) - pow(y, 2);
-    printf("(C code) Sampled: [%.2lf %.2lf] result %lf \n", x, y, t);
+float function(float x, float y) {
+    // float t = sin(x) + cos(y);
+    float t = -powf(x, 2) - powf(y, 2);
+    printf("(C code) Sampled: [%.2f %.2f] result %f \n", x, y, t);
     return t;
 }
 
-void learn(double *X_grid,
-           double *K,
-           double *L_T,
+void learn(float *X_grid,
+           float *K,
+           float *L_T,
            bool *sampled,
            int *X,
-           double *T,
+           float *T,
            int t,
            int maxIter,
-           double *mu,
-           double *sigma,
-           double(*kernel)(double *, double *, double *, double *),
-           const double beta,
+           float *mu,
+           float *sigma,
+           float(*kernel)(float *, float *, float *, float *),
+           const float beta,
            int n) {
-    /*
-     * grid_idx = self.argmax_ucb()
-    *  self.sample(self.X_grid[grid_idx])
-    *  for every point x:
-     *  gp_regression()
-    *  gp.fit(self.X, self.T)
-    *  mu1 = self.mu
-     */
     bool debug = true;
     int maxI = 0;
     int maxJ = 0;
-    double max = mu[0] + sqrt(beta) * sigma[0];
+    float max = mu[0] + sqrtf(beta) * sigma[0];
     for (int i = 0; i < n; i++) {
         for (int j = 0; j < n; j++) {
-            double currentValue = mu[i * n + j] + sqrt(beta) * sigma[i * n + j];
+            float currentValue = mu[i * n + j] + sqrtf(beta) * sigma[i * n + j];
 
             if (!sampled[i * n + j] && (currentValue > max)) {
                 max = currentValue;
@@ -109,10 +101,10 @@ void learn(double *X_grid,
                   n); // updating mu and sigma for every x in X_grid
 }
 
-double kernel2(double *x1, double *y1, double *x2, double *y2) {
+float kernel2(float *x1, float *y1, float *x2, float *y2) {
     // RBF kernel
-    double sigma = 1;
-    return exp(-((*x1 - *x2) * (*x1 - *x2) + (*y1 - *y2) * (*y1 - *y2)) / (2 * sigma * sigma));
+    float sigma = 1;
+    return expf(-((*x1 - *x2) * (*x1 - *x2) + (*y1 - *y2) * (*y1 - *y2)) / (2 * sigma * sigma));
 }
 
 void run() {
@@ -141,7 +133,7 @@ void clean() {
     n:    The size of the data in matrix A to decompose
     size: The actual size of the rows
  */
-void cholesky(double *A, int n, int size) {
+void cholesky(float *A, int n, int size) {
     for (int i = 0; i < n; ++i) {
 
         // Update the off diagonal entries first.
@@ -156,7 +148,7 @@ void cholesky(double *A, int n, int size) {
         for (int k = 0; k < i; ++k) {
             A[size * i + i] -= A[size * i + k] * A[size * i + k];
         }
-        A[size * i + i] = sqrt(A[size * i + i]);
+        A[size * i + i] = sqrtf(A[size * i + i]);
     }
 }
 
@@ -171,7 +163,7 @@ void cholesky(double *A, int n, int size) {
     size: The actual size of the rows
 
  */
-void incremental_cholesky(double *A, double *A_T, int n1, int n2, int size) {
+void incremental_cholesky(float *A, float *A_T, int n1, int n2, int size) {
     for (int i = n1; i < n2; ++i) {
         // Update the off diagonal entries.
         for (int j = 0; j < i; ++j) {
@@ -185,7 +177,7 @@ void incremental_cholesky(double *A, double *A_T, int n1, int n2, int size) {
         for (int k = 0; k < i; ++k) {
             A[size * i + i] -= A[size * i + k] * A[size * i + k];
         }
-        A[size * i + i] = sqrt(A[size * i + i]);
+        A[size * i + i] = sqrtf(A[size * i + i]);
         A_T[size * i + i] = A[size * i + i];
     }
 }
@@ -201,10 +193,10 @@ void incremental_cholesky(double *A, double *A_T, int n1, int n2, int size) {
  *      x: vector to put result in
  *      lower: if one the lower triangle system is solved, else the upper triangle system is solved.
 */
-void cholesky_solve2(int d, int size, double *LU, double *b, double *x, int lower) {
+void cholesky_solve2(int d, int size, float *LU, float *b, float *x, int lower) {
     if (lower == 1) {
         for (int i = 0; i < d; ++i) {
-            double sum = 0.;
+            float sum = 0.;
             for (int k = 0; k < i; ++k) {
                 sum += LU[i * size + k] * x[k];
             }
@@ -212,7 +204,7 @@ void cholesky_solve2(int d, int size, double *LU, double *b, double *x, int lowe
         }
     } else {
         for (int i = d - 1; i >= 0; --i) {
-            double sum = 0.;
+            float sum = 0.;
             for (int k = i + 1; k < d; ++k) {
                 sum += LU[i * size + k] * x[k];
             }
@@ -337,22 +329,22 @@ void cholesky_solve2_opt(int d, int size, float *LU, float *b, float *x, int low
 }
 
 // Old version.
-void cholesky_solve(int d, double *LU, double *b, double *x) {
-    double y[d];
+void cholesky_solve(int d, float *LU, float *b, float *x) {
+    float y[d];
     for (int i = 0; i < d; ++i) {
-        double sum = 0.;
+        float sum = 0.;
         for (int k = 0; k < i; ++k)sum += LU[i * d + k] * y[k];
         y[i] = (b[i] - sum) / LU[i * d + i];
     }
     for (int i = d - 1; i >= 0; --i) {
-        double sum = 0.;
+        float sum = 0.;
         for (int k = i + 1; k < d; ++k)sum += LU[k * d + i] * x[k];
         x[i] = (y[i] - sum) / LU[i * d + i];
     }
 }
 
 
-void transpose(double *M, double *M_T, int d, int size) {
+void transpose(float *M, float *M_T, int d, int size) {
     for (int i = 0; i < d; ++i) {
         for (int j = 0; j < d; ++j) {
             M_T[j * size + i] = M[i * size + j];
@@ -361,16 +353,16 @@ void transpose(double *M, double *M_T, int d, int size) {
 }
 
 
-void gp_regression(double *X_grid,
-                   double *K,
-                   double *L_T,
+void gp_regression(float *X_grid,
+                   float *K,
+                   float *L_T,
                    int *X,
-                   double *T,
+                   float *T,
                    int t,
                    int maxIter,
-                   double   (*kernel)(double *, double *, double *, double *),
-                   double *mu,
-                   double *sigma,
+                   float   (*kernel)(float *, float *, float *, float *),
+                   float *mu,
+                   float *sigma,
                    int n) {
     int t_gp = t + 1;
 
@@ -395,9 +387,9 @@ void gp_regression(double *X_grid,
     incremental_cholesky(K, L_T, t_gp - 1, t_gp, maxIter);
 
     // 3. Compute alpha
-    double x[t_gp];
-    double alpha[t_gp];
-    double v[t_gp];
+    float x[t_gp];
+    float alpha[t_gp];
+    float v[t_gp];
 
 
     cholesky_solve2(t_gp, maxIter, K, T, x, 1);
@@ -407,13 +399,13 @@ void gp_regression(double *X_grid,
 
     for (i = 0; i < n; i++) { // for all points in X_grid ([i])
         for (int j = 0; j < n; j++) { // for all points in X_grid ([i][j])
-            double x_star = X_grid[2 * n * i + 2 * j]; // Current grid point that we are looking at
-            double y_star = X_grid[2 * n * i + 2 * j + 1];
-            double k_star[t_gp];
-            double f_star = 0;
-            double variance = (*kernel)(&x_star, &y_star, &x_star, &y_star);
+            float x_star = X_grid[2 * n * i + 2 * j]; // Current grid point that we are looking at
+            float y_star = X_grid[2 * n * i + 2 * j + 1];
+            float k_star[t_gp];
+            float f_star = 0;
+            float variance = (*kernel)(&x_star, &y_star, &x_star, &y_star);
             int x_, y_;
-            double arg1x, arg1y, sum;
+            float arg1x, arg1y, sum;
 
             for (int k = 0; k < t_gp; k++) {
                 x_ = X[2 * k];
