+++ src/gpucb5.h
@@ -5,42 +5,42 @@
 
 #include <stdbool.h>
 
-double GRID_MIN_;
-double GRID_INC_;
-double BETA_;
+float GRID_MIN_;
+float GRID_INC_;
+float BETA_;
 extern const char *tag[10];
 // Allocate memory
-double *T_;
+float *T_;
 int *X_;
-double *X_grid_;
+float *X_grid_;
 bool *sampled_;
-double *mu_;
-double *sigma_;
-double *K_;
-double *L_;
+float *mu_;
+float *sigma_;
+float *K_;
+float *L_;
 
 int I_;
 int N_;
 
-double function(double x, double y);
+float function(float x, float y);
 
-void learn(double *X_grid,
-           double *K,
-           double *L_T,
+void learn(float *X_grid,
+           float *K,
+           float *L_T,
            bool *sampled,
            int *X,
-           double *T,
+           float *T,
            int t,
            int maxIter,
-           double *mu,
-           double *sigma,
-           double(*kernel)(double *, double *, double *, double *),
-           const double beta,
+           float *mu,
+           float *sigma,
+           float(*kernel)(float *, float *, float *, float *),
+           const float beta,
            int n);
 
-double kernel2(double *x1, double *y1, double *x2, double *y2);
+float kernel2(float *x1, float *y1, float *x2, float *y2);
 
-void initialize_meshgrid(double *X_grid, int n, double min, double inc);
+void initialize_meshgrid(float *X_grid, int n, float min, float inc);
 
 void initialize(const int, const int);
 
@@ -48,30 +48,30 @@ void run();
 
 void clean();
 
-int gpucb(int maxIter, int n, double grid_min, double grid_inc);
+int gpucb(int maxIter, int n, float grid_min, float grid_inc);
 
-void cholesky(double *A, int n, int size);
+void cholesky(float *A, int n, int size);
 
-void incremental_cholesky(double *A, double *A_T, int n1, int n2, int size);
+void incremental_cholesky(float *A, float *A_T, int n1, int n2, int size);
 
-void cholesky_solve2(int d, int size, double *LU, double *b, double *x, int lower);
+void cholesky_solve2(int d, int size, float *LU, float *b, float *x, int lower);
 
 void cholesky_solve2_opt(int d, int size, float *LU, float *b, float *x, int lower);
 
-void cholesky_solve(int d, double *LU, double *b, double *x);
+void cholesky_solve(int d, float *LU, float *b, float *x);
 
-void transpose(double *M, double *M_T, int d, int size);
+void transpose(float *M, float *M_T, int d, int size);
 
-void gp_regression(double *X_grid,
-                   double *K,
-                   double *L_T,
+void gp_regression(float *X_grid,
+                   float *K,
+                   float *L_T,
                    int *X,
-                   double *T,
+                   float *T,
                    int t,
                    int maxIter,
-                   double   (*kernel)(double *, double *, double *, double *),
-                   double *mu,
-                   double *sigma,
+                   float   (*kernel)(float *, float *, float *, float *),
+                   float *mu,
+                   float *sigma,
                    int n);
